refactor(codex): remove installer cleanup slop
