+++ src/cli/install-codex/install-codex.ts
@@ -1,7 +1,7 @@
 import { homedir } from "node:os"
-import { delimiter, join, resolve } from "node:path"
+import { join, resolve } from "node:path"
 import { existsSync } from "node:fs"
-import { mkdir, readFile, readlink, rm, stat, writeFile } from "node:fs/promises"
+import { mkdir, writeFile } from "node:fs/promises"
 import { installCachedPlugin, linkCachedPluginBins, pruneMarketplaceCache, pruneMarketplacePluginCaches } from "./codex-cache"
 import { shouldBuildSourcePackages } from "./codex-package-layout"
 import { updateCodexConfig } from "./codex-config-toml"
@@ -11,9 +11,12 @@ import { linkCachedPluginAgents } from "./link-cached-plugin-agents"
 import { readMarketplace, readPluginManifest, resolvePluginSource, validatePathSegment } from "./codex-marketplace"
 import { writeInstalledMarketplaceSnapshot, type MarketplaceSnapshotPluginSource } from "./codex-marketplace-snapshot"
 import { defaultRunCommand } from "./codex-process"
-import type { CodexInstallOptions, CodexInstallPlatform, CodexInstallResult, CodexMarketplaceSource, InstalledPlugin, MarketplaceManifest, OhMyCodexCleanupPrompt } from "./types"
+import { removeOhMyCodexBeforeInstall } from "./oh-my-codex-cleanup"
+import type { CodexInstallOptions, CodexInstallResult, CodexMarketplaceSource, InstalledPlugin, MarketplaceManifest } from "./types"
 
 const SISYPHUS_LEGACY_CACHE_MARKETPLACES = ["lazycodex", "code-yeongyu-codex-plugins"] as const
+
+export { removeOhMyCodexBeforeInstall } from "./oh-my-codex-cleanup"
 
 export async function runCodexInstaller(options: CodexInstallOptions = {}): Promise<CodexInstallResult> {
   const env = options.env ?? process.env
@@ -155,113 +158,6 @@ export async function runCodexInstaller(options: CodexInstallOptions = {}): Prom
     configPath,
     codexHome,
     gitBashPath: gitBashResolution.path,
-  }
-}
-
-export async function removeOhMyCodexBeforeInstall(input: {
-  readonly codexHome: string
-  readonly confirmCleanup?: OhMyCodexCleanupPrompt
-  readonly env: { readonly [key: string]: string | undefined }
-  readonly platform: CodexInstallPlatform
-  readonly repoRoot: string
-  readonly runCommand: typeof defaultRunCommand
-}): Promise<void> {
-  const omxPath = await findCommand("omx", input.env, input.platform)
-  let omxUninstallError: Error | null = null
-  if (omxPath && await isOhMyCodexCommand(omxPath)) {
-    if (input.confirmCleanup && !(await input.confirmCleanup({ omxPath }))) {
-      throw new Error("Codex install cancelled: existing oh-my-codex cleanup was not approved.")
-    }
-    try {
-      await input.runCommand("omx", ["uninstall", "--purge"], { cwd: input.repoRoot })
-    } catch (error) {
-      omxUninstallError = error instanceof Error ? error : new Error(String(error))
-    }
-  }
-
-  await input.runCommand("npm", ["uninstall", "-g", "oh-my-codex"], { cwd: input.repoRoot })
-  await removeOhMyCodexResidue(input.codexHome, input.repoRoot)
-
-  const remainingOmxPath = await findCommand("omx", input.env, input.platform)
-  if (remainingOmxPath && await isOhMyCodexCommand(remainingOmxPath)) {
-    await rm(remainingOmxPath, { force: true })
-  }
-
-  const verifiedOmxPath = await findCommand("omx", input.env, input.platform)
-  if (verifiedOmxPath && await isOhMyCodexCommand(verifiedOmxPath)) {
-    throw new Error(ohMyCodexCleanupFailureMessage(verifiedOmxPath, omxUninstallError))
-  }
-}
-
-function ohMyCodexCleanupFailureMessage(omxPath: string, uninstallError: Error | null): string {
-  const base = `oh-my-codex cleanup failed: omx is still installed at ${omxPath}`
-  return uninstallError ? `${base}; omx uninstall failed: ${uninstallError.message}` : base
-}
-
-async function isOhMyCodexCommand(path: string): Promise<boolean> {
-  if (isOhMyCodexPackagePath(path)) return true
-
-  try {
-    const target = await readlink(path)
-    if (isOhMyCodexPackagePath(target)) return true
-  } catch (error) {
-    if (!(error instanceof Error)) return false
-  }
-
-  try {
-    const content = await readFile(path, "utf8")
-    return isOhMyCodexShimContent(content)
-  } catch (error) {
-    if (error instanceof Error) return false
-    return false
-  }
-}
-
-function isOhMyCodexPackagePath(path: string): boolean {
-  return /(?:^|[\\/])node_modules[\\/]oh-my-codex(?:[\\/]|$)/.test(path) || /(?:^|[\\/])oh-my-codex[\\/]bin[\\/]omx(?:\.[^\\/]*)?$/.test(path)
-}
-
-function isOhMyCodexShimContent(content: string): boolean {
-  return /(?:^|[\\/])node_modules[\\/]oh-my-codex(?:[\\/]|$)/.test(content) || /_where=.*oh-my-codex/.test(content) || /require\(["']oh-my-codex(?:[\\/][^"']*)?["']\)/.test(content)
-}
-
-async function removeOhMyCodexResidue(codexHome: string, repoRoot: string): Promise<void> {
-  await rm(join(codexHome, "plugins", "cache", "oh-my-codex-local"), { recursive: true, force: true })
-  await rm(join(repoRoot, ".omx"), { recursive: true, force: true })
-}
-
-async function findCommand(
-  command: string,
-  env: { readonly [key: string]: string | undefined },
-  platform: CodexInstallPlatform,
-): Promise<string> {
-  const extensions = platform === "win32" ? (env.PATHEXT ?? ".EXE;.CMD;.BAT;.COM").split(";") : [""]
-  const names = extensions.map((extension) => {
-    const normalizedExtension = extension.toLowerCase()
-    const normalizedCommand = command.toLowerCase()
-    return normalizedCommand.endsWith(normalizedExtension) ? command : `${command}${extension}`
-  })
-
-  for (const directory of (env.PATH ?? "").split(delimiter)) {
-    if (directory.trim().length === 0) continue
-    for (const name of names) {
-      const path = join(directory, name)
-      if (await isCommandFile(path, platform)) return path
-    }
-  }
-
-  return ""
-}
-
-async function isCommandFile(path: string, platform: CodexInstallPlatform): Promise<boolean> {
-  try {
-    const fileStat = await stat(path)
-    if (!fileStat.isFile()) return false
-    if (platform === "win32") return true
-    return (fileStat.mode & 0o111) !== 0
-  } catch (error) {
-    if (error instanceof Error) return false
-    return false
   }
 }
